pyproject.toml
[project]
name = "todo"
version = "0.1.0"
description = "test app todo"
authors = [
    {name = "giveMepower94"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "django-allauth (>=65.12.0,<66.0.0)",
    "markdown (>=3.10,<4.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
